SHIN-21-Parts out Tic Tac Toe Game Logic & added socket namespaces
Game logic for the tic-tac-toe game was removed from the server.js file and placed in the tictactoe_server.js file. The tictactoe_server function from this new module was then called from the server.js file. 
Namespaces were also added into the server.js file to segreate the communications of the chatroom and the game.
